portfolio_projects/data_engineering/stock_market_analysis/requirements.txt
# Apache Airflow
apache-airflow==2.10.0
apache-airflow-providers-apache-kafka==3.1.0
apache-airflow-providers-apache-spark==4.0.0
apache-airflow-providers-cassandra==2.4.0
apache-airflow-providers-postgres==5.1.0
apache-airflow-providers-http==4.6.0
apache-airflow-providers-ftp==3.6.0

# Spark & Kafka
pyspark==3.4.1
kafka-python==2.0.2

# Cassandra
cassandra-driver==3.28.0

# Additional dependencies
aiohttp==3.8.5
alembic==1.12.0
anyio==4.0.0
apispec==6.3.0
argcomplete==3.1.1
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
backoff==2.2.1
blinker==1.6.2
cachelib==0.9.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
dill==0.3.7
dnspython==2.4.2
docutils==0.20.1
email-validator==1.3.1
Flask==2.2.5
Flask-AppBuilder==4.3.3
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
geomet==0.2.1.post1
googleapis-common-protos==1.60.0
grpcio==1.57.0
gunicorn==21.2.0
httpx==0.24.1
idna==3.4
importlib-resources==6.0.1
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.0
lazy-object-proxy==1.9.0
limits==3.6.0
lockfile==0.12.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
opentelemetry-api==1.15.0
opentelemetry-sdk==1.15.0
ordered-set==4.1.0
packaging==23.1
pathspec==0.11.2
pendulum==2.1.2
pluggy==1.3.0
prison==0.2.1
protobuf==4.24.2
psutil==5.9.5
python-daemon==3.0.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rich==13.5.2
setproctitle==1.3.2
six==1.16.0
SQLAlchemy==1.4.49
tabulate==0.9.0
tenacity==8.2.3
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
yarl==1.9.2


